:truck: move: 폴더, 파일 이동 및 분리 (#40)
- server/api 폴더 생성
- 그 밑으로 기존 pages/api에 있던 파일들 옮김
- auth.ts 같은 경우는 external api와 supabase api를 분리하여 작성
