+++ src/hooks/table/useSetTable.ts
@@ -1,4 +1,4 @@
-import { addStoreTable } from '@/pages/api/store-table';
+import { addStoreTable } from '@/server/api/supabase/store-table';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useEffect, useRef } from 'react';
