--- src/pages/api/hello.ts
@@ -1,1 +1,0 @@
-// Hello world API
